python312_qt6/docstyle-pro_claude_v4/requirements.txt
python-docx>=1.1.0
lxml>=5.0.0
PyQt6>=6.0.0
Pillow>=10.0.0
google-auth-oauthlib
google-api-python-client
chromadb
openai
anthropic
google-genai
markdown
onnxruntime<=1.19.2
